refactor: simplify module structure
